Support new gotchiverse coordinate system and parcel name search
This is now displayed in game under the minimap, and coordinates are
relative to the top-left corner of the district i.e. need to know both
the district and the x, y offset within it.
When on a parcel, the parcel name is displayed in-game, so also allow
searching by name in the same place.